+++ TP3/geneticwarrior/src/main/java/interfaces/Phenotype.java
@@ -5,5 +5,5 @@ public interface Phenotype {
 
     String toString();
 
-    Double getValue(String key);
+    Float getValue(String key);
 }
